# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Aplicação Spring Boot REST API
  version: 1.0.0
  title: Gestão de Veículos
  contact:
    name: Infnet
    url: https://www.vr.com.br/
    email: infnet@edu.com.br
  license:
    name: Apache License Version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
host: localhost:8989
basePath: /
...
paths:
...
  /veiculos/all/by/ids:
    get:
      tags:
        - veiculo-controller
      summary: Serviço responsável por recuperar veículos pelos identificadores
      operationId: getVeiculosByIdsUsingGET
      produces:
        - '*/*'
      parameters:
        - name: ids
          in: query
          description: ids
          required: true
          type: array
          items:
            type: integer
            format: int32
          collectionFormat: multi
      responses:
        '200':
          description: Veículos recuperado com sucesso!
          schema:
            type: array
            items:
              $ref: '#/definitions/Veiculo'
        '404':
          description: Serviço indisponível!
  /veiculos/veiculo/{id}/excluir:
    get:
      tags:
        - veiculo-controller
      summary: Serviço responsável por excluir o veiculo pelo identificador
      operationId: excluirUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Veículo excluído com sucesso!
          schema:
            type: string
        '404':
          description: Serviço indisponível!
definitions:
  Veiculo:
    type: object
    properties:
      id:
        type: integer
        format: int32
      marca:
        type: string
      nome:
        type: string
      valor:
        type: number
        format: float
    title: Veiculo